+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
-- `DatabaseCredentials` in favor of `SqlAlchemyConnector` - [#35](https://github.com/PrefectHQ/prefect-sqlalchemy/pull/35)
+- `DatabaseCredentials` and all tasks in `prefect_sqlalchemy.database`, in favor of `SqlAlchemyConnector` - [#35](https://github.com/PrefectHQ/prefect-sqlalchemy/pull/35)
 
 ### Removed